Policies, Images and Validations

+++ database/migrations/2022_05_11_141055_create_operative_systems_table.php
@@ -16,7 +16,7 @@ class CreateOperativeSystemsTable extends Migration
         Schema::create('operative_systems', function (Blueprint $table) {
             $table->id();
             $table->string('name', 15);
-            $table->integer('version');
+            $table->float('version');
         });
     }
